+++ src/item_type.h
@@ -270,8 +270,8 @@ public:
   int32_t runeMagLevel = 0;
   int32_t runeLevel = 0;
 
-  uint16_t id = 0;
-  uint16_t clientId = 0;
+  uint32_t id = 0;
+  uint32_t clientId = 0;
 
   // CombatType_t combatType = COMBAT_NONE;
   uint16_t rotateTo = 0;
