+++ src/map_view.h
@@ -215,7 +215,7 @@ private:
 
 	void selectRegion(const Position &from, const Position &to);
 	void removeItemsInRegion(const Position &from, const Position &to, std::function<bool(const Item &)> predicate);
-	void fillRegion(const Position &from, const Position &to, uint16_t serverId);
+	void fillRegion(const Position &from, const Position &to, uint32_t serverId);
 
 	void cameraViewportChangedEvent();
 };
